requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator>=2.2.0
python-dotenv==1.0.1
aiofiles==24.1.0 